+++ src/WiredSlider.tsx
@@ -1,7 +1,7 @@
-// TODO: knobRadius and zero color are not behaving correctly. See storybook.
 import * as React from 'react';
 import { BaseProps } from './types';
 import { useCustomElement } from './utils/useCustomElement';
+import { debounce } from './utils/debounce';
 const { useMemo } = React;
 
 export interface WiredSliderProps extends BaseProps {
@@ -20,11 +20,6 @@ export interface WiredSliderProps extends BaseProps {
    * @default 100
    */
   max?: number;
-  /**
-   *  The radius of the knob.
-   * @default 100
-   */
-  knobRadius?: number;
   /**
    *  Color of the knob when the value is at minimum.
    * @default black
@@ -51,7 +46,6 @@ export const WiredSlider = ({
   value = 30,
   min = 0,
   max = 100,
-  knobRadius,
   knobZeroColor = 'black',
   knobColor = 'rgba(0, 0, 200, 0.8)',
   barColor = 'currentColor',
@@ -59,24 +53,15 @@ export const WiredSlider = ({
 }: WiredSliderProps) => {
   const customValues = useMemo(() => {
     return {
-      attributes: { value, min, max, knobradius: knobRadius },
+      attributes: { value, min, max },
       css: {
-        '--wired-slider-knob-zero-color': knobZeroColor,
-        '--wired-slider-knob-color': knobColor,
+        // '--wired-slider-knob-zero-color': knobZeroColor,
+        '--wired-slider-knob-color': value === min ? knobZeroColor : knobColor,
         '--wired-slider-bar-color': barColor,
       },
-      methods: { change: onChange },
+      methods: { change: debounce(onChange, 500) },
     };
-  }, [
-    value,
-    min,
-    max,
-    knobRadius,
-    knobZeroColor,
-    knobColor,
-    barColor,
-    onChange,
-  ]);
+  }, [value, min, max, knobZeroColor, knobColor, barColor, onChange]);
 
   const register = useCustomElement(customValues);
   return <wired-slider ref={register}></wired-slider>;