+++ src/utils/debounce.ts
@@ -1,0 +1,16 @@
+export function debounce(func: any, wait: number, immediate?: boolean) {
+  let timeout: any;
+  return function() {
+    // @ts-ignore
+    var context = this,
+      args = arguments;
+    var later = function() {
+      timeout = null;
+      if (!immediate) func.apply(context, args);
+    };
+    var callNow = immediate && !timeout;
+    clearTimeout(timeout);
+    timeout = setTimeout(later, wait);
+    if (callNow) func.apply(context, args);
+  };
+}