+++ stories/WiredSlider.stories.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { WiredSlider } from '../src';
 import { color, number } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
@@ -15,17 +15,23 @@ export default {
   title: 'WiredSlider',
 };
 
-export const Default = () => (
-  <Container>
-    <WiredSlider
-      value={0}
-      min={number('min', 0, options)}
-      max={number('max', 100, options)}
-      knobRadius={number('knobRadius', 10)}
-      knobZeroColor={color('knobZeroColor', 'black')}
-      knobColor={color('knobColor', 'rgba(0, 0, 200, 0.8)')}
-      barColor={color('barColor', 'currentColor')}
-      onChange={action('onChange Fired')}
-    />
-  </Container>
-);
+export const Default = () => {
+  const [value, setValue] = useState(30);
+  function handleChange(e) {
+    setValue(e.target.value);
+    action('onChange Fired', e);
+  }
+  return (
+    <Container>
+      <WiredSlider
+        value={value}
+        min={number('min', 0, options)}
+        max={number('max', 100, options)}
+        knobZeroColor={color('knobZeroColor', 'black')}
+        knobColor={color('knobColor', 'rgba(0, 0, 200, 0.8)')}
+        barColor={color('barColor', 'currentColor')}
+        onChange={handleChange}
+      />
+    </Container>
+  );
+};